fixes for Numba 0.53.0dev0
It seems like previous Numba versions (up to 0.51.2) may have worked 'by
accident'. When running pynndescent against a recent development version
of Numba. I receievd many errors during test collection of the form:

```
E   numba.core.errors.LoweringError: Failed in nopython mode pipeline (step: nopython mode backend)
E   Storing i64 to ptr of i32 ('dim'). FE type uint32
E
E   File "../miniconda3/envs/umap/lib/python3.7/site-packages/pynndescent/utils.py", line 88:
E   def norm(vec):
E       <source elided>
E       result = 0.0
E       dim = vec.shape[0]
E       ^
E
E   During: lowering "dim = static_getitem(value=$8load_attr.2, index=0, index_var=$const10.3, fn=<built-in function getitem>)" at /Users/vhaenel/git/numba-integration-testing/miniconda3/envs/umap/lib/python3.7/site-packages/pynndescent/utils.py (88)
```

Not all errors are shown immediately as pytest fails on the first
encounter, so they have to be fixed one-by-one.

+++ pynndescent/rp_trees.py
@@ -853,7 +853,7 @@ def make_sparse_tree(inds, indptr, spdata, rng_state, leaf_size=30, angular=Fals
     fastmath=True,
     locals={
         "margin": numba.types.float32,
-        "dim": numba.types.uint16,
+        "dim": numba.types.int64,
         "d": numba.types.uint16,
     },
 )
